Define multi-stage commands as subcommands
The multi-stage args used by install and upgrade were implemented as
positional arguments to their parent commands. This made the
documentation unclear, and the code ambiguous as to which flags
corresponded to which stage.

Define `config` and `control-plane` stages as subcommands. The help
menus now explicitly state flags supported.

Fixes #2729

Signed-off-by: Andrew Seigner <siggy@buoyant.io>